replace request with requestV2 refactor
